Manufacturing costing — review fixes: split monitor Playwright into its own spec + grid-fix timestamp
- Move the cost-imbalance-monitor Playwright test out of mfg-config-translations.spec.js
  (which had an unrelated translations-only setup) into its own mfg-cost-imbalance-monitor.spec.js,
  per the one-setup-per-spec-file rule. Both specs pass locally.
- Correct the grid-order migration Updated timestamp so it no longer predates the row Created.
