+++ backend/de.metas.adempiere.adempiere/migration/src/main/sql/postgresql/system/10-de.metas.adempiere/5814720_sys_gh30811_PP_Order_CostImbalance_Window_GridOrder.sql
@@ -6,6 +6,6 @@
 -- (SeqNo) and the other columns are unchanged. AD_UI_Element has no _Trl and no propagation
 -- dependency, so a direct UPDATE is correct.
 UPDATE AD_UI_Element SET SeqNoGrid=35,
-       Updated=TO_TIMESTAMP('2026-07-20 10:10:00','YYYY-MM-DD HH24:MI:SS'), UpdatedBy=100
+       Updated=TO_TIMESTAMP('2026-07-20 23:49:54','YYYY-MM-DD HH24:MI:SS'), UpdatedBy=100
 WHERE AD_UI_Element_ID=652684
 ;
